requirements/requirements.txt
# requirements for the production project.

requests==2.32.5
websocket-client==1.8.0
